Add animated player sprite

+++ src/lib.rs
@@ -3,7 +3,7 @@ pub struct Starfield {
     pub pos_x: f32,
     pub pos_y: f32,
     pub speed: f32,
-    pub color: f32,
+    pub color: u8,
 }
 
 #[cfg(test)]
